extracted RestTimeSelect
